Update start.sh
every time we run this bash file all dependencies must be check for install but know if we have dependencies or installed. don't reinstalling all of them and run code quickly

+++ start.sh
@@ -35,12 +35,14 @@ fi
 VPY="$VENV_DIR/bin/python"
 
 echo "[*] Installing dependencies ..."
+if ! "$VPY" -m pip install -r requirements.txt -q --dry-run; then
 "$VPY" -m pip install --disable-pip-version-check -q --upgrade pip >/dev/null
 if ! "$VPY" -m pip install --disable-pip-version-check -q -r requirements.txt; then
     echo "[!] PyPI install failed. Retrying via runflare mirror ..."
     "$VPY" -m pip install --disable-pip-version-check -q -r requirements.txt \
         -i https://mirror-pypi.runflare.com/simple/ \
         --trusted-host mirror-pypi.runflare.com
+fi
 fi
 
 if [ ! -f "config.json" ]; then
